✅ #78 : AlertManager Class 생성, 공용으로 사용하는 lyricsAlert 함수 생성
