+++ 가사오케/GasaOK/Controller/SongFolder/ViewController.swift
@@ -143,7 +143,8 @@ extension ViewController: UITableViewDelegate, UITableViewDataSource {
         guard let singer = index.value(forKey: "singer") as? String else { return }
         
         /// 노래제목과 가수를 파라미터로 함수 호출
-        showLyricsAlert(title: songTitle, singer: singer)
+        AlertManager.shared.lyricsAlert(vc: self, title: songTitle, singer: singer)
+
     }
     
     // MARK: - 데이터 fetch
@@ -171,26 +172,4 @@ extension ViewController: UITableViewDelegate, UITableViewDataSource {
         }
     }
     
-    // MARK: - 가사 보기 알림창
-    ///FIXME: 변수명 제안 lyricsAlertWillShow
-    func showLyricsAlert(title: String, singer: String) {
-        let alert = UIAlertController(title: "가사를 보시겠습니까?", message: "가사 저작권에 의해 앱 내에서 바로 가사를 보여드릴 수 없습니다.\n링크를 통해 가사를 확인하시겠습니까?", preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: "취소", style: .cancel))
-        alert.addAction(UIAlertAction(title: "이동", style: .default, handler: { _ in
-            let baseURL = "https://m.search.naver.com/search.naver?sm=mtp_hty.top&where=m&query="
-            var titleArray: [Character] = []
-            for char in title{
-                if char == "(" { break }
-                titleArray.append(char)
-                
-            }
-            let realTitle = titleArray.map{String($0)}.joined()
-            print(realTitle)
-            let url = baseURL + realTitle.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed)! + "+" + singer.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed)! + "+" + "가사".addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed)!
-            let searchURL = URL(string: url)
-            UIApplication.shared.open(searchURL!, options: [:])
-        }))
-        
-        self.present(alert, animated: false)
-    }
 }
